+++ web/tour/checkframe.js
@@ -1,3 +1,3 @@
 // redirects page if not in iframe
 
-if (top === self) location.href = window.location.origin + '/?T=tour&U=' + parseInt(window.location.href.match(/tour(\d+)/)[1]);
+if (top === self) window.location.href = window.location.origin + '/?T=tour&U=' + parseInt(window.location.href.match(/tour(\d+)/)[1]);